% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLTPFunctions.R
\name{apply_test_control_model}
\alias{apply_test_control_model}
\title{Apply a linear model and t-test to a dataset}
\usage{
apply_test_control_model(dataset)
}
\arguments{
\item{dataset}{A dataframe containing the data for analysis. The dataset must include at least the columns "normalised_amplitude" and "Pathway".}
}
\value{
The result of a t-test applied to the model formula <normalised_amplitude ~ Pathway>.
}
\description{
This function takes a dataset, standardizes the "Pathway" column as a factor, fits a linear model using the formula \code{normalised_amplitude ~ Pathway}, and performs a t-test on the same formula. It returns the results of the t-test.
}
