% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_test_compare.R
\name{filter_five_bin}
\alias{filter_five_bin}
\title{Filter Dataset by Time Range}
\usage{
filter_five_bin(dataset, bin_start)
}
\arguments{
\item{dataset}{A dataframe containing time-series data.}

\item{bin_start}{The start of the 5-minute time range.}
}
\value{
A filtered dataframe containing data within the specified time range.
}
\description{
This function filters a dataset to include only rows within a 5-minute bin starting from the provided bin start time.
}
