% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_unique_cells.R
\name{find_unique_cells}
\alias{find_unique_cells}
\title{Get a list of unique combined animal and cell IDs}
\usage{
find_unique_cells(dataset)
}
\arguments{
\item{dataset}{A dataframe containing the column \code{animal_cell_id}, which holds the combined animal and cell IDs.}
}
\value{
A vector of unique \code{animal_cell_id} values from the dataset.
}
\description{
This function extracts the unique values from the \code{animal_cell_id} column of the given dataset and returns them as a vector. The \code{animal_cell_id} typically represents a combined identifier for both animal and cell.
}
