% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLTPFunctions.R
\name{generate_bin_title}
\alias{generate_bin_title}
\title{Generate Bin Title}
\usage{
generate_bin_title(binstart, binend)
}
\arguments{
\item{binstart}{The starting value of the bin.}

\item{binend}{The ending value of the bin.}
}
\value{
A string representing the title for the bin range.
}
\description{
This function generates a title based on the start and end of a binning range.
}
