% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_test_compare.R
\name{generate_full_bin_figure_function}
\alias{generate_full_bin_figure_function}
\title{Generate Full Bin Plot}
\usage{
generate_full_bin_figure_function(dataset, bin_start = 20, legend = TRUE)
}
\arguments{
\item{dataset}{A dataframe containing time-series data.}

\item{bin_start}{The start of the time bin for plotting. Defaults to 20.}

\item{legend}{Logical, whether to include the legend in the plot. Defaults to TRUE.}
}
\value{
A ggplot object visualizing the data within the specified bin.
}
\description{
This function generates a full plot for a specified time bin by filtering the dataset and generating a plot using the 5-minute bin plotting function.
}
