% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_functions.R
\name{generate_means_from_bins}
\alias{generate_means_from_bins}
\title{Generate Mean Values from Bins}
\usage{
generate_means_from_bins(dataset)
}
\arguments{
\item{dataset}{A dataframe containing test and control pathway data.}
}
\value{
A long-form dataframe containing the mean values for test and control pathways.
}
\description{
This function calculates the mean values for test and control pathways, grouping by animal or cell ID, and transforms the dataset into long format.
}
