% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longform_function.R
\name{longform_test_control}
\alias{longform_test_control}
\title{Reshape dataset into long format for test and control pathways}
\usage{
longform_test_control(dataset, is_mean = FALSE)
}
\arguments{
\item{dataset}{A dataframe containing "Test", "Control", and optionally SEM columns.}

\item{is_mean}{Logical; if \code{TRUE}, the function includes SEM columns in the output.}
}
\value{
A long-format dataframe with pathways and amplitudes.
}
\description{
This function reshapes the dataset into long format, with "Test" and "Control" values as separate pathways, and optionally includes standard error of the mean (SEM) columns.
}
