% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLTPFunctions.R
\name{longitudinal_and_binplot}
\alias{longitudinal_and_binplot}
\title{Combine Longitudinal and Bin Plots}
\usage{
longitudinal_and_binplot(longplot, binplot)
}
\arguments{
\item{longplot}{A ggplot object representing the longitudinal plot.}

\item{binplot}{A ggplot object representing the bin plot.}
}
\value{
A combined plot object with the longitudinal and bin plots side by side.
}
\description{
This function arranges a longitudinal plot and a bin plot side by side for comparison.
}
