% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ltppackage-package}
\alias{ltppackage}
\alias{ltppackage-package}
\title{ltp_analysis}
\description{
Produces diagnostic and results plots for ltp data
}
\author{
\strong{Maintainer}: Peter Richardson \email{richardsonp2@cardiff.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
