% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_run_functions.R
\name{main_combined_run}
\alias{main_combined_run}
\title{Main Combined Run}
\usage{
main_combined_run(dataset)
}
\arguments{
\item{dataset}{A dataframe containing data for test and control pathways.}
}
\value{
A ggplot object visualizing the mean and SEM for test and control pathways.
}
\description{
This function processes a dataset by cleaning outliers, calculating mean and SEM values, transforming the dataset into long-form format, and generating a plot for test and control pathways.
}
