% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_run_functions.R
\name{main_run}
\alias{main_run}
\title{Generate and save diagnostic plots into a PowerPoint slide deck}
\usage{
main_run(dataset, individual_value, ppt)
}
\arguments{
\item{dataset}{A dataframe containing the data, which should include the column "animal_cell_id".}

\item{individual_value}{A value used to filter the dataset to include only the data for a specific individual (e.g., animal or cell).}

\item{ppt}{A PowerPoint object created using the \code{officer} package, where the slides will be added.}
}
\value{
The updated PowerPoint object with the new slide containing the title and combined diagnostic plots.
}
\description{
This function processes a dataset, generates diagnostic and test/control plots, and saves them as a slide in a PowerPoint deck. The function filters the dataset based on a specific individual value, preprocesses the data, handles outliers, and arranges the plots into a single figure. The final figure is added to a PowerPoint presentation as a new slide.
}
