% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLTPFunctions.R
\name{make_diagnostic_grid}
\alias{make_diagnostic_grid}
\title{Create a diagnostic grid of Rs, Rin, and DC figures}
\usage{
make_diagnostic_grid(dataset)
}
\arguments{
\item{dataset}{A dataframe that contains the numeric columns "Rs", "Rin", and "DC".}
}
\value{
A grid of diagnostic plots arranged horizontally.
}
\description{
This function generates diagnostic plots for three parameters: "Rs", "Rin", and "DC", and arranges them into a grid layout with one row and three columns.
}
