% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosistics.R
\name{make_diagnostic_plot}
\alias{make_diagnostic_plot}
\title{The whole dataset is preprocessed here. Then the filtered dataset is filtered according to the cells I pick
after manually choosing the right ones.
Title of the function}
\usage{
make_diagnostic_plot(dataset, parameter)
}
\arguments{
\item{x}{Description of the input parameter.}
}
\value{
What the function returns.
}
\description{
Description of what the function does.
}
