% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLTPFunctions.R
\name{make_test_control_plot}
\alias{make_test_control_plot}
\title{Create an error bar plot with mean and SEM for test and control pathways}
\usage{
make_test_control_plot(dataset)
}
\arguments{
\item{dataset}{A dataframe containing mean and SEM values for "Test" and "Control" pathways.}
}
\value{
A ggplot object for visualizing the mean and SEM for test and control pathways.
}
\description{
This function generates a plot with error bars to visualize the mean and standard error of the mean (SEM) for the "Test" and "Control" pathways over time.
}
