% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosistics.R
\name{make_test_control_plot_single}
\alias{make_test_control_plot_single}
\title{Create a single diagnostic plot for test and control pathways}
\usage{
make_test_control_plot_single(dataset)
}
\arguments{
\item{dataset}{A long-format dataframe containing "Test" and "Control" values for different time points.}
}
\value{
A ggplot object for visualizing test and control pathways.
}
\description{
This function generates a plot for diagnostic purposes using normalized amplitude values for "Test" and "Control" pathways over time.
}
