% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLTPFunctions.R
\name{plot_fivemin_function}
\alias{plot_fivemin_function}
\title{Plot 5-Minute Bin Data}
\usage{
plot_fivemin_function(dataset, title, legend = TRUE)
}
\arguments{
\item{dataset}{A dataframe containing binned data.}

\item{title}{The title for the plot.}

\item{legend}{Logical, whether to include the legend in the plot. Defaults to TRUE.}
}
\value{
A ggplot object visualizing the binned data.
}
\description{
This function generates a box plot and jitter plot to visualize normalized amplitudes for test and control pathways within a 5-minute bin.
}
