% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLTPFunctions.R
\name{pre_process_datasets}
\alias{pre_process_datasets}
\title{Preprocess the dataset by standardizing column names and converting data types}
\usage{
pre_process_datasets(dataset)
}
\arguments{
\item{dataset}{A dataframe containing the data to be preprocessed. The columns should correspond to the expected variables (e.g., "Key", "Date", "Sex", "GT", "animal_id", "cell_num", etc.).}
}
\value{
A dataframe with standardized column names and appropriately converted data types (factors and numeric).
}
\description{
This function takes a dataset, standardizes its column names, and converts specific columns to the appropriate data types (factors or numeric). It ensures that certain columns are treated as factors (e.g., categorical variables) and others as numeric. This step prepares the dataset for further analysis.
}
