% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv.R
\name{read_csv_function}
\alias{read_csv_function}
\title{Read LTP CSV files}
\usage{
read_csv_function(address)
}
\arguments{
\item{address}{The string address for the csv file.}
}
\value{
A dataframe containing the 15 columns.
}
\description{
A narrow focus read csv function which will import data, correct headings and check if the dataframe contains exactly 15 columns.
Not having the 15 columns will result in issues.
}
