% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing_functions.R
\name{replace_outliers}
\alias{replace_outliers}
\title{Replace outliers in the dataset with the mean value}
\usage{
replace_outliers(dataset)
}
\arguments{
\item{dataset}{A dataframe containing numeric columns.}
}
\value{
A dataframe with outliers replaced by the mean of the corresponding column.
}
\description{
This function checks each numeric column (except "Time") for outliers (values that deviate more than 2 standard deviations from the mean) and replaces them with the mean value of the non-zero "Time" data.
}
