% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLTPFunctions.R
\name{run_main_diagnostic}
\alias{run_main_diagnostic}
\title{Run the main diagnostic plot generation}
\usage{
run_main_diagnostic(dataset)
}
\arguments{
\item{dataset}{A dataframe containing the necessary data for generating diagnostic plots.}
}
\value{
A combined plot with a title and a grid of diagnostic plots.
}
\description{
This function combines a title and a grid of diagnostic plots into a single plot, and prepares it for visualization.
}
