% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosistics.R
\name{theme_blank_background}
\alias{theme_blank_background}
\title{Custom theme with transparent background for plots}
\usage{
theme_blank_background(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{Base font size.}

\item{base_family}{Base font family.}
}
\value{
A ggplot2 theme object with customized settings.
}
\description{
This function provides a minimal ggplot2 theme with a transparent background and customized axis lines, tick marks, and legend formatting.
}
