% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_test_control_plot_functions.R
\name{make_test_control_plot_single}
\alias{make_test_control_plot_single}
\title{Create a single diagnostic plot for test and control pathways}
\usage{
make_test_control_plot_single(
  dataset,
  top_amplitude_limit = 3,
  right_timepoint_max_limit = 35
)
}
\arguments{
\item{dataset}{A long-format dataframe containing "Test" and "Control" values for different time points.}

\item{top_amplitude_limit}{The highest amplitude to plot up to. Higher than 3 will stretch the plot.}

\item{right_timepoint_max_limit}{The furthest timepoint value from 0 to the right to plot on the figure.}
}
\value{
A ggplot object for visualizing test and control pathways.
}
\description{
This function generates a plot for diagnostic purposes using normalized amplitude values for "Test" and "Control" pathways over time.
}
