% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerpoint_generation.R
\name{print_try_function}
\alias{print_try_function}
\title{Run PowerPoint generation for a given dataset and handle errors}
\usage{
print_try_function(ppt, cell_id, dataframe)
}
\arguments{
\item{ppt}{A PowerPoint object created using the \code{officer} package, to which new slides will be added.}

\item{cell_id}{The identifier for the specific cell or individual being processed.}

\item{dataframe}{The dataset being processed, which will be either a HET or WT dataset.}
}
\value{
The updated PowerPoint object if successful, or the original PowerPoint object if an error occurs.
}
\description{
This function attempts to generate and insert slides into a PowerPoint presentation based on a specific \code{cell_id} and a given dataset. It checks if the dataset contains het or wt and prints a message about the dataset being processed. The function handles errors during slide generation without crashing the entire process. If an error occurs, it logs the error and returns the PowerPoint object unchanged.
}
