% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BaseLTPFunctions.R
\name{read_csv_function}
\alias{read_csv_function}
\title{Read LTP CSV files}
\usage{
read_csv_function(address)
}
\arguments{
\item{address}{The string address for the csv file.}
}
\value{
A dataframe containing the 15 columns.
}
\description{
Description of what the function does.
}
